Flytt ut journalpost til egen fil
